refactor(model): remove non-model entry points
