+++ spikingjelly/activation_based/model/parametric_lif_net.py
@@ -1,4 +1,3 @@
-from spikingjelly.logger import logger
 from copy import deepcopy
 
 import torch
@@ -298,75 +297,3 @@ class DVSGestureNet(nn.Module):
 
     def forward(self, x: torch.Tensor):
         return self.conv_fc(x)
-
-
-def test_models():
-    import torch
-
-    from .. import functional, neuron, surrogate
-
-    x = torch.rand([2, 1, 28, 28])
-    net = MNISTNet(16, neuron.IFNode, surrogate_function=surrogate.ATan())
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    functional.set_step_mode(net, "m")
-    x = torch.rand([4, 2, 1, 28, 28])
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    del net
-    del x
-
-    x = torch.rand([2, 1, 28, 28])
-    net = FashionMNISTNet(16, neuron.IFNode, surrogate_function=surrogate.ATan())
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    functional.set_step_mode(net, "m")
-    x = torch.rand([4, 2, 1, 28, 28])
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    del net
-    del x
-
-    x = torch.rand([2, 2, 32, 32])
-    net = NMNISTNet(16, neuron.IFNode, surrogate_function=surrogate.ATan())
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    functional.set_step_mode(net, "m")
-    x = torch.rand([4, 2, 2, 32, 32])
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    del net
-    del x
-
-    x = torch.rand([2, 3, 32, 32])
-    net = CIFAR10Net(16, neuron.IFNode, surrogate_function=surrogate.ATan())
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    functional.set_step_mode(net, "m")
-    x = torch.rand([4, 2, 3, 32, 32])
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    del net
-    del x
-
-    x = torch.rand([2, 2, 128, 128])
-    net = CIFAR10DVSNet(16, neuron.IFNode, surrogate_function=surrogate.ATan())
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    functional.set_step_mode(net, "m")
-    x = torch.rand([4, 2, 2, 128, 128])
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    del net
-    del x
-
-    x = torch.rand([2, 2, 128, 128])
-    net = DVSGestureNet(16, neuron.IFNode, surrogate_function=surrogate.ATan())
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    functional.set_step_mode(net, "m")
-    x = torch.rand([4, 2, 2, 128, 128])
-    logger.info("{}", net(x).shape)
-    functional.reset_net(net)
-    del net
-    del x
