+++ spikingjelly/activation_based/model/spikformer.py
@@ -6,11 +6,6 @@ from ..layer.attention import SpikingSelfAttention
 
 __all__ = [
     "Spikformer",
-    "SpikformerBlock",
-    "SpikformerConv2dBN",
-    "SpikformerConv2dBNLIF",
-    "SpikformerMLP",
-    "SpikformerPatchStem",
     "spikformer_cifar10",
     "spikformer_s",
     "spikformer_ti",
